+++ workspaces/azure-devops/plugins/azure-devops/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-azure-devops
 
+## 0.24.0
+
+### Minor Changes
+
+- f9618c9: Backstage version bump to v1.46.2
+
+### Patch Changes
+
+- Updated dependencies [f9618c9]
+  - @backstage-community/plugin-azure-devops-common@0.18.0
+
 ## 0.23.0
 
 ### Minor Changes