+++ workspaces/jenkins/plugins/jenkins-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-jenkins-backend
 
+## 0.23.0
+
+### Minor Changes
+
+- 95d7628: Backstage version bump to v1.46.2
+
+### Patch Changes
+
+- Updated dependencies [95d7628]
+  - @backstage-community/plugin-jenkins-common@0.15.0
+
 ## 0.22.0
 
 ### Minor Changes