+++ workspaces/jenkins/plugins/scaffolder-backend-module-jenkins/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-scaffolder-backend-module-jenkins
 
+## 0.17.0
+
+### Minor Changes
+
+- 95d7628: Backstage version bump to v1.46.2
+
+### Patch Changes
+
+- Updated dependencies [95d7628]
+  - @backstage-community/plugin-jenkins-common@0.15.0
+
 ## 0.16.0
 
 ### Minor Changes